requirements.txt
duckdb>=1.0.0
numpy>=1.25.2
pandas>=2.2.2
matplotlib>=3.8.4
gradio>=4.44.0
langgraph>=0.2.34
google-generativeai>=0.8.3
python-dotenv>=1.0.1
grandalf>=0.8